+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This file tries to follow recommendations of [Keep a CHANGELOG](http://keepachan
 ### Added
 - Account Create (Signup Express)
 - Application Read & Find
+- Application Create
 
 ## [0.1.3] - 2016-05-09
 ### Added